Update solidity-1.md
The condition in SetDougAddress function in DougEnabled didn't make sense to me. Perhaps I have misunderstood but currently, assuming DOUG != 0x0, the new dougAddr must be equal to DOUG in order to proceed past the conditional statement? In which case, the statement following, DOUG = dougAddr would be redundant as they are the same anyway and consequently it is not possible for any contract, not even the current DOUG, to change the address of DOUG?

Checking for msg.sender != DOUG made sense to me?